refactor: shortform 관련 기능 패키지 통합
