+++ src/main/java/com/ssafy/tiggle/service/shortform/news/NewsCrawlerServiceImpl.java
@@ -1,7 +1,7 @@
-package com.ssafy.tiggle.service.news;
+package com.ssafy.tiggle.service.shortform.news;
 
-import com.ssafy.tiggle.dto.news.CategoryNewsResponseDto;
-import com.ssafy.tiggle.dto.news.NewsArticleDto;
+import com.ssafy.tiggle.dto.shortform.news.CategoryNewsResponseDto;
+import com.ssafy.tiggle.dto.shortform.news.NewsArticleDto;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
